#5 refactor : common 패키지 구조를 리팩토링 한다
