+++ src/main/java/snowflake/tumblog/common/BaseException.java
@@ -3,6 +3,7 @@ package snowflake.tumblog.common;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
+import snowflake.tumblog.common.enums.BaseResponseStatus;
 
 @Getter
 @Setter
