+++ src/main/java/snowflake/tumblog/common/constants/Constant.java
@@ -1,9 +1,8 @@
-package snowflake.tumblog.common;
+package snowflake.tumblog.common.constants;
 
 public class Constant {
+
     public static final String ACTIVE = "active";
     public static final String INACTIVE = "inactive";
-    private final static String baseURI = "/api/v1";
-    public final static String userURI = baseURI + "/users";
     public final static Integer ZERO = 0;
 }
