+++ src/main/java/snowflake/tumblog/common/constants/RequestURI.java
@@ -1,0 +1,10 @@
+package snowflake.tumblog.common.constants;
+
+public class RequestURI {
+
+    private final static String baseURI = "/api/v1";
+
+    public final static String user = baseURI + "/users";
+    public final static String ocr = baseURI + "/ocr";
+
+}
